cmd: move tikv-server and tikv-ctl to cmd crate (#5242)
