+++ Cargo.toml
@@ -25,12 +25,6 @@ failpoints = ["fail/failpoints"]
 [lib]
 name = "tikv"
 
-[[bin]]
-name = "tikv-server"
-
-[[bin]]
-name = "tikv-ctl"
-
 [[test]]
 name = "failpoints"
 path = "tests/failpoints/mod.rs"
@@ -87,7 +81,6 @@ utime = "0.2"
 chrono = "0.4"
 lazy_static = "1.3"
 backtrace = "0.2.3"
-clap = "2.32"
 url = "1.7.2"
 sys-info = "0.5.7"
 futures = "0.1"
@@ -195,6 +188,7 @@ members = [
   "components/tidb_query_datatype",
   "components/tidb_query_codegen",
   "components/pd_client",
+  "cmd",
 ]
 
 [profile.dev]
