+++ cmd/src/lib.rs
@@ -1,0 +1,11 @@
+// Copyright 2019 TiKV Project Authors. Licensed under Apache-2.0.
+
+#[macro_use(slog_error, slog_warn, slog_info, slog_crit)]
+extern crate slog;
+#[macro_use]
+extern crate slog_global;
+
+#[macro_use]
+pub mod setup;
+pub mod server;
+pub mod signal_handler;
