+++ cmd/src/server.rs
@@ -2,26 +2,7 @@
 
 use super::setup::*;
 use super::signal_handler;
-use crate::binutil::setup::initial_logger;
-use crate::config::{check_and_persist_critical_config, TiKvConfig};
-use crate::coprocessor;
-use crate::fatal;
-use crate::import::{ImportSSTService, SSTImporter};
-use crate::raftstore::coprocessor::{CoprocessorHost, RegionInfoAccessor};
-use crate::raftstore::store::fsm::store::{StoreMeta, PENDING_VOTES_CAP};
-use crate::raftstore::store::PdTask;
-use crate::raftstore::store::{fsm, LocalReader};
-use crate::raftstore::store::{new_compaction_listener, SnapManagerBuilder};
-use crate::server::resolve;
-use crate::server::status_server::StatusServer;
-use crate::server::transport::ServerRaftStoreRouter;
-use crate::server::DEFAULT_CLUSTER_ID;
-use crate::server::{create_raft_storage, Node, RaftKv, Server};
-use crate::storage::lock_manager::{
-    Detector, DetectorScheduler, Service as DeadlockService, WaiterManager, WaiterMgrScheduler,
-};
-use crate::storage::{self, AutoGCConfig, DEFAULT_ROCKSDB_SUB_DIR};
-use crate::storage::{FlowStatistics, FlowStatsReporter};
+use crate::setup::initial_logger;
 use engine::rocks;
 use engine::rocks::util::metrics_flusher::{MetricsFlusher, DEFAULT_FLUSHER_INTERVAL};
 use engine::rocks::util::security::encrypted_env_from_cipher_file;
@@ -33,11 +14,26 @@ use std::path::Path;
 use std::sync::{Arc, Mutex};
 use std::thread::JoinHandle;
 use std::time::Duration;
+use tikv::config::{check_and_persist_critical_config, TiKvConfig};
+use tikv::coprocessor;
+use tikv::import::{ImportSSTService, SSTImporter};
+use tikv::raftstore::coprocessor::{CoprocessorHost, RegionInfoAccessor};
+use tikv::raftstore::store::fsm::store::{StoreMeta, PENDING_VOTES_CAP};
+use tikv::raftstore::store::{fsm, LocalReader};
+use tikv::raftstore::store::{new_compaction_listener, SnapManagerBuilder};
+use tikv::server::resolve;
+use tikv::server::status_server::StatusServer;
+use tikv::server::transport::ServerRaftStoreRouter;
+use tikv::server::DEFAULT_CLUSTER_ID;
+use tikv::server::{create_raft_storage, Node, RaftKv, Server};
+use tikv::storage::lock_manager::{
+    Detector, DetectorScheduler, Service as DeadlockService, WaiterManager, WaiterMgrScheduler,
+};
+use tikv::storage::{self, AutoGCConfig, DEFAULT_ROCKSDB_SUB_DIR};
 use tikv_util::check_environment_variables;
-use tikv_util::collections::HashMap;
 use tikv_util::security::SecurityManager;
 use tikv_util::time::Monitor;
-use tikv_util::worker::{FutureScheduler, FutureWorker};
+use tikv_util::worker::FutureWorker;
 
 const RESERVED_OPEN_FDS: u64 = 1000;
 
@@ -53,7 +49,7 @@ pub fn run_tikv(mut config: TiKvConfig) {
     tikv_util::set_panic_hook(false, &config.storage.data_dir);
 
     // Print version information.
-    super::log_tikv_info();
+    tikv::log_tikv_info();
 
     config.compatible_adjust();
     if let Err(e) = config.validate() {
@@ -88,14 +84,6 @@ pub fn run_tikv(mut config: TiKvConfig) {
 
     let _m = Monitor::default();
     run_raft_server(pd_client, &config, security_mgr);
-}
-
-impl FlowStatsReporter for FutureScheduler<PdTask> {
-    fn report_read_stats(&self, read_stats: HashMap<u64, FlowStatistics>) {
-        if let Err(e) = self.schedule(PdTask::ReadStats { read_stats }) {
-            error!("Failed to send read flow statistics"; "err" => ?e);
-        }
-    }
 }
 
 fn run_raft_server(pd_client: RpcClient, cfg: &TiKvConfig, security_mgr: Arc<SecurityManager>) {
