+++ cmd/src/setup.rs
@@ -6,7 +6,7 @@ use std::sync::atomic::{AtomicBool, Ordering};
 use chrono;
 use clap::ArgMatches;
 
-use crate::config::{MetricConfig, TiKvConfig};
+use tikv::config::{MetricConfig, TiKvConfig};
 use tikv_util::collections::HashMap;
 use tikv_util::{self, logger};
 
@@ -16,7 +16,7 @@ pub static LOG_INITIALIZED: AtomicBool = AtomicBool::new(false);
 #[macro_export]
 macro_rules! fatal {
     ($lvl:expr, $($arg:tt)+) => ({
-        if $crate::binutil::setup::LOG_INITIALIZED.load(::std::sync::atomic::Ordering::SeqCst) {
+        if $crate::setup::LOG_INITIALIZED.load(::std::sync::atomic::Ordering::SeqCst) {
             crit!($lvl, $($arg)+);
         } else {
             eprintln!($lvl, $($arg)+);
