+++ src/lib.rs
@@ -38,7 +38,7 @@ extern crate prometheus;
 extern crate quick_error;
 #[macro_use]
 extern crate serde_derive;
-#[macro_use(slog_trace, slog_error, slog_warn, slog_info, slog_debug, slog_crit)]
+#[macro_use(slog_trace, slog_error, slog_warn, slog_info, slog_debug)]
 extern crate slog;
 #[macro_use]
 extern crate slog_derive;
@@ -58,10 +58,34 @@ extern crate failure;
 #[cfg(test)]
 extern crate test;
 
-pub mod binutil;
 pub mod config;
 pub mod coprocessor;
 pub mod import;
 pub mod raftstore;
 pub mod server;
 pub mod storage;
+
+/// Returns the tikv version information.
+pub fn tikv_version_info() -> String {
+    let fallback = "Unknown (env var does not exist when building)";
+    format!(
+        "\nRelease Version:   {}\
+         \nGit Commit Hash:   {}\
+         \nGit Commit Branch: {}\
+         \nUTC Build Time:    {}\
+         \nRust Version:      {}",
+        env!("CARGO_PKG_VERSION"),
+        option_env!("TIKV_BUILD_GIT_HASH").unwrap_or(fallback),
+        option_env!("TIKV_BUILD_GIT_BRANCH").unwrap_or(fallback),
+        option_env!("TIKV_BUILD_TIME").unwrap_or(fallback),
+        option_env!("TIKV_BUILD_RUSTC_VERSION").unwrap_or(fallback),
+    )
+}
+
+/// Prints the tikv version information to the standard output.
+pub fn log_tikv_info() {
+    info!("Welcome to TiKV");
+    for line in tikv_version_info().lines() {
+        info!("{}", line);
+    }
+}
